Rewrite connection logic using local orthonormal bases
The structure of the bases corresponds with Minecraft's coordinate system. Since y is used as the up/down dimension, the actual structure of the bases is tangent-normal-bitangent, instead of the typical standard. Regardless, this change to the codebase brings it to my desired quality level for future expansion. Basing all calculations purely from a math standpoint as such makes a lot of room for accesssible implementation of improvements down the line.